get bearerToken from Authorization header, get token after splitting the bearerToken, treat error case for a wrong formatted bearerToken and for an invalid token

+++ Rest/src/main/java/controller/ExpenseController.java
@@ -1,6 +1,7 @@
 package controller;
 
 import domain.ExpenseCategory;
+import domain.User;
 import dto.ExpenseDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.format.annotation.DateTimeFormat;
@@ -13,6 +14,7 @@ import service.exception.ServiceException;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Map;
+import java.util.Optional;
 
 @CrossOrigin
 @RestController
@@ -31,33 +33,59 @@ public class ExpenseController {
     }
 
 
-    @DeleteMapping("/delete-expense/{expenseId}/{userId}")
-    public ResponseEntity<?> delete(@PathVariable int expenseId, @PathVariable int userId) {
-        //TODO get userId from token; remove userId param
-        try {
-            return new ResponseEntity<>(service.deleteExpense(expenseId, userId), HttpStatus.OK);
-        } catch (ServiceException e) {
-            if (e.getMessage().equals("Forbidden access to this expense"))
-                return new ResponseEntity<>(e.getMessage(), HttpStatus.FORBIDDEN);
-            else if (e.getMessage().equals("Internal server error"))
-                return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
-            else
-                return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
+    @DeleteMapping("/delete-expense/{expenseId}")
+    public ResponseEntity<?> delete(@PathVariable int expenseId, @RequestHeader("Authorization") String bearerToken) {
+
+        if (bearerToken != null && bearerToken.startsWith("Bearer")) {
+            String token = bearerToken.substring(7);
+            Optional<User> user = service.getTokenUser(token);
+            if (user.isPresent()) {
+                int userId = user.get().getId();
+                try {
+                    return new ResponseEntity<>(service.deleteExpense(expenseId, userId), HttpStatus.OK);
+                } catch (ServiceException e) {
+                    if (e.getMessage().equals("Forbidden access to this expense"))
+                        return new ResponseEntity<>(e.getMessage(), HttpStatus.FORBIDDEN);
+                    else if (e.getMessage().equals("Internal server error"))
+                        return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+                    else
+                        return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
+                }
+            }
+            else{
+                return new ResponseEntity<>("Unauthorized",HttpStatus.UNAUTHORIZED);
+            }
+        }
+        else{
+            return new ResponseEntity<>("Forbidden",HttpStatus.FORBIDDEN);
         }
     }
 
     @GetMapping("/get-expenses")
     public ResponseEntity<?> getExpenses(
-            @RequestParam int userId,
+            @RequestHeader("Authorization") String bearerToken,
             @RequestParam String category,
             @RequestParam long startDate,
             @RequestParam long endDate
     ) {
-        //TODO get userId from token; remove userId param
-        try {
-            return new ResponseEntity<>(service.getExpenses(userId, category, startDate, endDate), HttpStatus.OK);
-        } catch (ServiceException e) {
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
+
+        if (bearerToken != null && bearerToken.startsWith("Bearer")) {
+            String token = bearerToken.substring(7);
+            Optional<User> user = service.getTokenUser(token);
+            if (user.isPresent()) {
+                int userId = user.get().getId();
+                try {
+                    return new ResponseEntity<>(service.getExpenses(userId, category, startDate, endDate), HttpStatus.OK);
+                } catch (ServiceException e) {
+                    return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
+                }
+            }
+            else{
+                return new ResponseEntity<>("Unauthorized",HttpStatus.UNAUTHORIZED);
+            }
+        }
+        else{
+            return new ResponseEntity<>("Forbidden",HttpStatus.FORBIDDEN);
         }
 
     }
@@ -73,23 +101,51 @@ public class ExpenseController {
 
     @GetMapping("/total-expenses-in-time")
     public ResponseEntity<?> getTotalExpensesInTime(
-            @RequestParam int userId,
+            @RequestHeader("Authorization") String bearerToken,
             @RequestParam String granularity,
             @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate startDate,
             @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate endDate,
             @RequestParam String category) {
-        try {
-            return new ResponseEntity<>(service.getTotalExpensesInTime(userId, granularity, startDate, endDate, category), HttpStatus.OK);
-        } catch (ServiceException e) {
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
+
+        if (bearerToken != null && bearerToken.startsWith("Bearer")) {
+            String token = bearerToken.substring(7);
+            Optional<User> user = service.getTokenUser(token);
+            if (user.isPresent()) {
+                int userId = user.get().getId();
+                try {
+                    return new ResponseEntity<>(service.getTotalExpensesInTime(userId, granularity, startDate, endDate, category), HttpStatus.OK);
+                } catch (ServiceException e) {
+                    return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
+                }
+            }
+            else{
+                return new ResponseEntity<>("Unauthorized",HttpStatus.UNAUTHORIZED);
+            }
+        }
+        else{
+            return new ResponseEntity<>("Forbidden",HttpStatus.FORBIDDEN);
         }
     }
 
     @GetMapping(value = "/category-total")
-    public ResponseEntity<?> getExpensesTotalByCategory(@RequestParam("userId") int userId,
+    public ResponseEntity<?> getExpensesTotalByCategory(@RequestHeader("Authorization") String bearerToken,
                                                         @RequestParam("start") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime start,
                                                         @RequestParam("end") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime end) {
-        Map<ExpenseCategory, Double> result = service.getExpenseTotalByCategory(userId, start, end);
-        return new ResponseEntity<>(result, HttpStatus.OK);
+
+        if (bearerToken != null && bearerToken.startsWith("Bearer")) {
+            String token = bearerToken.substring(7);
+            Optional<User> user = service.getTokenUser(token);
+            if (user.isPresent()) {
+                int userId = user.get().getId();
+                Map<ExpenseCategory, Double> result = service.getExpenseTotalByCategory(userId, start, end);
+                return new ResponseEntity<>(result, HttpStatus.OK);
+            }
+            else{
+                return new ResponseEntity<>("Unauthorized",HttpStatus.UNAUTHORIZED);
+            }
+        }
+        else{
+            return new ResponseEntity<>("Forbidden",HttpStatus.FORBIDDEN);
+        }
     }
 }
